+++ GloryEngine/Engine/GloryCore/RendererModule.h
@@ -23,7 +23,7 @@ namespace Glory
 		void OnGameThreadFrameStart();
 		void OnGameThreadFrameEnd();
 
-		virtual RenderTexture* CreateCameraRenderTexture(size_t width, size_t height);
+		virtual RenderTexture* CreateCameraRenderTexture(uint32_t width, uint32_t height);
 		virtual void GetCameraRenderTextureAttachments(std::vector<Attachment>& atachments);
 		virtual void OnCameraResize(CameraRef camera);
 
@@ -42,7 +42,7 @@ namespace Glory
 		virtual void Initialize() override;
 		virtual void Cleanup() = 0;
 		virtual void OnRender(CameraRef camera, const RenderData& renderData, const std::vector<PointLight>& lights = std::vector<PointLight>()) = 0;
-		virtual void OnDoScreenRender(CameraRef camera, const FrameData<PointLight>& lights, size_t width, size_t height, RenderTexture* pRenderTexture) = 0;
+		virtual void OnDoScreenRender(CameraRef camera, const FrameData<PointLight>& lights, uint32_t width, uint32_t height, RenderTexture* pRenderTexture) = 0;
 
 		virtual void OnInitialize() {};
 		virtual void OnThreadedInitialize() {}
